src/bot/pyproject.toml
[project]
name = "newsagent"
version = "0.1.0"
description = "Telegram бот для сбора и отправки новостей"
requires-python = ">=3.8"
dependencies = [
    "python-telegram-bot>=20.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

